made appropriate changes

+++ server/app/Http/Controllers/Api/V1/ReportController.php
@@ -7,66 +7,45 @@ use App\Http\Requests\StoreReportRequest;
 use App\Http\Resources\ReportResource;
 use App\Models\Report;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ReportController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function index()
     {
-        return ReportResource::collection(Report::with('reporter')->paginate(10));
+        $user = request()->user();
+        $reports = $user->reports()->paginate();
+        return $reports;
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(StoreReportRequest $request)
     {
         $data = $request->validated();
-        $data['user_id'] = 23;
+        $data['user_id'] = $request->user()->id;
+
         $report = Report::create($data);
         return new ReportResource($report);
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function show(Report $report)
     {
+        abort_if(Auth::id() != $report->reporter_id, 403, 'Access Forbidden!');
         return new ReportResource($report);
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function update(StoreReportRequest $request, Report $report)
     {
+        abort_if(Auth::id() != $report->reporter_id, 403, 'Access Forbidden!');
+
         $data = $request->validated();
         $report->update($data);
         return new ReportResource($report);
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function destroy(Report $report)
     {
+        abort_if(Auth::id() != $report->reporter_id, 403, 'Access Forbidden!');
+
         $report->delete();
         return response()->noContent();
     }
